Implemented two new effects, CooldownManager that keeps track of cooldown of effects and check for cooldown on EffectService

+++ src/main/kotlin/git/juampa99/rpgfy/healthbar/listener/HealthBarListener.kt
@@ -1,12 +1,9 @@
-package git.juampa99.rpgfy.healthbar.event
+package git.juampa99.rpgfy.healthbar.listener
 
-import git.juampa99.rpgfy.Rpgfy
 import git.juampa99.rpgfy.gear.effect.event.EffectFadeEvent
 import git.juampa99.rpgfy.gear.effect.event.EffectTriggerEvent
 import git.juampa99.rpgfy.healthbar.service.HealthBarService
 import git.juampa99.rpgfy.healthbar.service.HealthBarService.updateHealthBar
-import net.minecraft.server.v1_16_R3.MinecraftServer
-import org.bukkit.Bukkit
 import org.bukkit.Material
 import org.bukkit.entity.Creature
 import org.bukkit.event.EventHandler
@@ -42,13 +39,15 @@ class HealthBarListener : Listener {
 
     @EventHandler
     fun onEffectTrigger(event: EffectTriggerEvent) {
-        HealthBarService.addEffectToEntity(event.entity, event.effect)
+        if(event.effect.isDebuff())
+            HealthBarService.addEffectToEntity(event.entity, event.effect)
         updateHealthBar(event.entity)
     }
 
     @EventHandler
     fun onEffectFade(event: EffectFadeEvent) {
-        HealthBarService.removeEffectOfEntity(event.entity, event.effect)
+        if(event.effect.isDebuff())
+            HealthBarService.removeEffectOfEntity(event.entity, event.effect)
         updateHealthBar(event.entity)
     }
